[MOD] Renamed module and now related fields
