--- mrp_manufacturing_order_sale_line/__openerp__.py
@@ -1,39 +1,0 @@
-# -*- encoding: utf-8 -*-
-##############################################################################
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see http://www.gnu.org/licenses/.
-#
-##############################################################################
-{
-    "name": "MRP Manufacturing Order Sale Line",
-    "version": "1.0",
-    "author": "OdooMRP team",
-    "category": "MRP",
-    "website": "http://www.odoomrp.com",
-    "description": """
-    This module performs the following:
-
-    In the production order two new fields are created: customer, sale line.
-
-    when the production order is create, these two new fields are loaded with
-    sale order information.
-    """,
-    "depends": ['sale',
-                'mrp',
-                'procurement',
-                ],
-    "data": ['views/mrp_production_view.xml',
-             ],
-    "installable": True
-}
