--- mrp_manufacturing_order_sale_line/i18n/mrp_manufacturing_order_sale_line.pot
@@ -1,42 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* mrp_manufacturing_order_sale_line
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 8.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2014-10-16 11:37+0000\n"
-"PO-Revision-Date: 2014-10-16 11:37+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: mrp_manufacturing_order_sale_line
-#: view:mrp.production:mrp_manufacturing_order_sale_line.view_mrp_production_filter_inh_withsaleinfo
-msgid "Customer"
-msgstr ""
-
-#. module: mrp_manufacturing_order_sale_line
-#: model:ir.model,name:mrp_manufacturing_order_sale_line.model_mrp_production
-msgid "Manufacturing Order"
-msgstr ""
-
-#. module: mrp_manufacturing_order_sale_line
-#: model:ir.model,name:mrp_manufacturing_order_sale_line.model_procurement_order
-msgid "Procurement"
-msgstr ""
-
-#. module: mrp_manufacturing_order_sale_line
-#: view:mrp.production:mrp_manufacturing_order_sale_line.view_mrp_production_filter_inh_withsaleinfo
-msgid "Product"
-msgstr ""
-
-#. module: mrp_manufacturing_order_sale_line
-#: field:mrp.production,sale_line:0
-msgid "Sale Line"
-msgstr ""
-
