--- mrp_manufacturing_order_sale_line/models/mrp_production.py
@@ -1,26 +1,0 @@
-# -*- encoding: utf-8 -*-
-##############################################################################
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see http://www.gnu.org/licenses/.
-#
-##############################################################################
-from openerp import models, fields
-
-
-class MrpProduction(models.Model):
-    _inherit = 'mrp.production'
-
-    partner = fields.Many2one('res.partner', string='Customer', store=True,
-                              related='sale_line.order_id.partner_id')
-    sale_line = fields.Many2one('sale.order.line', string='Sale Line')
