--- mrp_manufacturing_order_sale_line/views/mrp_production_view.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <record id="mrp_production_tree_view_inh_withsaleinfo" model="ir.ui.view">
-            <field name="name">mrp.production.tree.view.inh.withsaleinfo</field>
-            <field name="model">mrp.production</field>
-            <field name="inherit_id" ref="mrp.mrp_production_tree_view"/>
-            <field name="arch" type="xml">
-                <field name="name" position="after">
-                    <field name="partner" />
-                </field>
-            </field>
-        </record>
-        <record id="mrp_production_form_view_inh_withsaleinfo" model="ir.ui.view">
-            <field name="name">mrp.production.form.view.inh.withsaleinfo</field>
-            <field name="model">mrp.production</field>
-            <field name="inherit_id" ref="mrp.mrp_production_form_view"/>
-            <field name="arch" type="xml">
-                <xpath expr="//form/sheet/group/group/field[@name='product_id']" position="before">
-                    <field name="partner" domain="[('customer','=',True)]" attrs="{'invisible':[('partner','=',False)]}"/>
-                </xpath>
-                <xpath expr="//form/sheet/group/group/field[@name='bom_id']" position="before">
-                    <field name="sale_line" readonly="1" attrs="{'invisible': [('sale_line', '=', False)]}"/>
-                </xpath>
-            </field>
-        </record>
-        
-        
-        <record id="view_mrp_production_filter_inh_withsaleinfo" model="ir.ui.view">
-            <field name="name">view.mrp.production.filter.inh.withsaleinfo</field>
-            <field name="model">mrp.production</field>
-            <field name="inherit_id" ref="mrp.view_mrp_production_filter"/>
-            <field name="arch" type="xml">
-                <field name="product_id" position="before">
-                    <field name="partner" />
-                </field>
-                <filter string="Product" position="before">
-                    <filter string="Customer" icon="terp-accessories-archiver" domain="[]" context="{'group_by':'partner'}"/>
-                </filter>
-            </field>
-        </record>
-    </data>
-</openerp>
