+++ mrp_production_sale_info/README.rst
@@ -1,0 +1,4 @@
+MRP - Production sale info
+==========================
+
+This module show sale order related info in manufacturing orders.
