+++ mrp_production_sale_info/i18n/es.po
@@ -1,0 +1,39 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* mrp_production_sale_info
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 8.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2015-01-23 11:25+0000\n"
+"PO-Revision-Date: 2015-01-23 11:25+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: mrp_production_sale_info
+#: view:mrp.production:mrp_production_sale_info.mrp_production_sale_info_search_view
+#: field:mrp.production,partner:0
+msgid "Customer"
+msgstr "Cliente"
+
+#. module: mrp_production_sale_info
+#: model:ir.model,name:mrp_production_sale_info.model_mrp_production
+msgid "Manufacturing Order"
+msgstr "Órden de producción"
+
+#. module: mrp_production_sale_info
+#: view:mrp.production:mrp_production_sale_info.mrp_production_sale_info_search_view
+msgid "Product"
+msgstr "Producto"
+
+#. module: mrp_production_sale_info
+#: view:mrp.production:mrp_production_sale_info.mrp_production_sale_info_search_view
+#: field:mrp.production,sale_order:0
+msgid "Sale Order"
+msgstr "Pedido de venta"
+
