+++ mrp_production_sale_info/models/__init__.py
@@ -1,0 +1,6 @@
+# -*- encoding: utf-8 -*-
+##############################################################################
+# For copyright and license notices, see __openerp__.py file in root directory
+##############################################################################
+
+from . import mrp_production
